Added project for Xcode, which is the standard IDE for development on OS X 10.3 (Panther).  This project builds static libraries for G3D and GLG3D.
I had to make a minor change to glh/glut.h so it could
find glu.h on OS X.

There were also a couple of changes to the various Makefile.OSX
files.  The Makefile way of building the G3D libraries
on OS X is being superseded by Xcode, but it can't hurt
to keep them up to date for a little while longer.