Various updates to the Xcode project for the OS X build of G3D.
- Added a separate target for the Independent JPEG Group files.
This now builds to a static library, libjpeg.a, which is placed
in the same location as libG3D.a and libGLG3D.a

- Added SDL files necessary to write SDL applications for OS X.  These
files, SDLMain.h and SDLMain.m, must be contained in all SDL apps for
OS X or these apps will not link.  SDLMain.h and SDLMain.m are part of
the standard distribution of SDL 1.2 for OS X, and they're part of the
Project Builder SDL project templates. There are no project templates
for Xcode yet, so I'm including these files in the G3DOSX
subdirectory.  Note that the ZeroLink feature on Xcode must be
disabled for any SDL apps to run.

- Added a new target, SDL_Test.  This is a simple SDL application
just to demonstrate that SDL applications can be linked and
run properly.  This has been a major stumbling block on the OS X
port of G3D, so this test is relatively important.

- Do not use the header glx.h anymore for OS X. glx.h is the
X11 header for OpenGL use, and we don't want to have to use
the X server with G3D on OS X.

- Added an implementation for glGetProcAddress for OS X.
I got this from Apple tech note QA11888.  The implementation is
in the inline function glGetProcAddress in glcalls.h.